requirements.txt
asgiref==3.4.1
Django==3.2.8
django-annoying==0.10.6
django-filter==21.1
django-jsonstore==0.5.0
django-material==1.9.0
django-viewflow==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-ldap==3.3.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
